requirements.txt
fastapi
uvicorn
python-dotenv
twilio
google-generativeai
elevenlabs